Coordinate Translation, Text UI, Keyboard flying
Coordinates are now translated to a range of [-1,1] for easier
processing.

Detected objects will now be briefly displayed with a TUI (blessed).

The Drone will now navigate towards the mean of the detected objects.

We have very basic flight controls via keyboard. Use space to takeoff
and land, WASD to move forward and back, turn.